#25 스프링 데이터 JPA - MemberRepository 기존 클래스를 JpaRepository 상속받은 인터페이스화하여 사용. - findOne() -> findById() 변경 - findByName() 메서드 추가
