+++ src/main/java/com/junshock/jpatest/repository/MemberRepositoryOld.java
@@ -1,0 +1,35 @@
+package com.junshock.jpatest.repository;
+
+import com.junshock.jpatest.domain.Member;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+import javax.persistence.EntityManager;
+import java.util.List;
+
+@Repository
+@RequiredArgsConstructor
+public class MemberRepositoryOld {
+
+    private final EntityManager em;
+
+    public Long save(Member member) {
+        em.persist(member); // db insert를 이시점에서 안함. db전략 gericvalue에서는. db commit 될대 insert한다.
+        return member.getId();
+    }
+
+    public Member findOne(Long id) {
+        return em.find(Member.class, id);
+    }
+
+    public List<Member> findAll() {
+        return em.createQuery("select m from Member m", Member.class) // JSQL 테이블이 아닌 엔터티 객체를 대상으로 질의
+                .getResultList();
+    }
+
+    public List<Member> findByName(String name) {
+        return em.createQuery("select m from Member m where m.name = :name", Member.class)
+                .setParameter("name", name)
+                .getResultList();
+    }
+}
